update
fixed interactions with Stitcher and FileWalker.

There is currently an error with the Reader not finding path of the file
its trying to create (need to add file creation if found path doesn't
exist)

Writer class needs to be updated to write to the same location as reader
class.  Not sure if it will need file creation capabilities, since
Reader should be creating the file.